Moved controllers to their own submodule
