+++ autogator/routines.py
@@ -26,235 +26,10 @@ import keyboard
 from pydantic import BaseModel, BaseSettings
 
 from autogator.hardware import DataAcquisitionUnitBase, Stage
+from autogator.controllers import KeyboardControl
 
 
 log = logging.getLogger(__name__)
-
-
-class KeyloopKeyboardBindings(BaseSettings):
-    MOVE_LEFT: str = "left arrow"
-    MOVE_RIGHT: str = "right arrow"
-    MOVE_UP: str = "up arrow"
-    MOVE_DOWN: str = "down arrow"
-    JOG_LEFT: str = "a"
-    JOG_RIGHT: str = "d"
-    JOG_UP: str = "w"
-    JOG_DOWN: str = "s"
-    JOG_CLOCKWISE: str = "c"
-    JOG_COUNTERCLOCKWISE: str = "x"
-    LINEAR_JOG_STEP: str = "shift + j"
-    ROTATIONAL_JOG_STEP: str = "shift + g"
-    STOP_ALL: str = "space"
-    HOME: str = "o"
-    HELP: str = "h"
-    QUIT: str = "q"
-
-
-class KeyboardControl:
-    def __init__(self, stage: Stage, bindings: KeyloopKeyboardBindings = None):
-        self.stage = stage
-
-        if bindings is None:
-            bindings = KeyloopKeyboardBindings()
-        self.bindings = bindings
-
-        self.semaphores = {
-            "MOTOR_X" : threading.Semaphore(),
-            "MOTOR_Y" : threading.Semaphore(),
-            "MOTOR_PSI" : threading.Semaphore(),
-        }
-
-        self.linear_step_size = 0.1
-        self.rotational_step_size = 0.1
-
-    def _move_left(self):
-        semaphore = self.semaphores["MOTOR_X"]
-        if semaphore.acquire(timeout=0.1):
-            self.stage.x.move_cont("backward")
-            while keyboard.is_pressed(self.bindings.MOVE_LEFT):
-                time.sleep(0.05)
-            self.stage.x.stop()
-            semaphore.release()
-
-    def _move_right(self):
-        semaphore = self.semaphores["MOTOR_X"]
-        if semaphore.acquire(timeout=0.1):
-            self.stage.x.move_cont("forward")
-            while keyboard.is_pressed(self.bindings.MOVE_RIGHT):
-                time.sleep(0.05)
-            self.stage.x.stop()
-            semaphore.release()
-
-    def _move_up(self):
-        semaphore = self.semaphores["MOTOR_Y"]
-        if semaphore.acquire(timeout=0.1):
-            self.stage.y.move_cont("forward")
-            while keyboard.is_pressed(self.bindings.MOVE_UP):
-                time.sleep(0.05)
-            self.stage.y.stop()
-            semaphore.release()
-
-    def _move_down(self):
-        semaphore = self.semaphores["MOTOR_Y"]
-        if semaphore.acquire(timeout=0.1):
-            self.stage.y.move_cont("backward")
-            while keyboard.is_pressed(self.bindings.MOVE_DOWN):
-                time.sleep(0.05)
-            self.stage.y.stop()
-            semaphore.release()
-
-    def _jog_left(self):
-        semaphore = self.semaphores["MOTOR_X"]
-        if semaphore.acquire(timeout=0.1):
-            self.stage.x.move_by(-self.linear_step_size)
-            semaphore.release()
-
-    def _jog_right(self):
-        semaphore = self.semaphores["MOTOR_X"]
-        if semaphore.acquire(timeout=0.1):
-            self.stage.x.move_by(self.linear_step_size)
-            semaphore.release()
-
-    def _jog_up(self):
-        semaphore = self.semaphores["MOTOR_Y"]
-        if semaphore.acquire(timeout=0.1):
-            self.stage.y.move_by(self.linear_step_size)
-            semaphore.release()
-
-    def _jog_down(self):
-        semaphore = self.semaphores["MOTOR_Y"]
-        if semaphore.acquire(timeout=0.1):
-            self.stage.y.move_by(-self.linear_step_size)
-            semaphore.release()
-
-    def _jog_cw(self):
-        semaphore = self.semaphores["MOTOR_PSI"]
-        if semaphore.acquire(timeout=0.1):
-            self.stage.psi.move_by(self.rotational_step_size)
-            semaphore.release()
-
-    def _jog_ccw(self):
-        semaphore = self.semaphores["MOTOR_PSI"]
-        if semaphore.acquire(timeout=0.1):
-            self.stage.psi.move_by(-self.rotational_step_size)
-            semaphore.release()
-
-    def _set_linear_jog_step(self):
-        val = None
-        while val is None:
-            answer = input(f"Enter new step size or [ENTER] to cancel (current {self.linear_step_size}): ")
-            if answer == "n":
-                return
-            try:
-                val = float(answer)
-            except ValueError:
-                pass
-        self.linear_step_size = val
-        print(f"New step size set: {self.linear_step_size}")
-
-    def _set_rotational_jog_step(self):
-        val = None
-        while val is None:
-            answer = input(f"Enter new step size or [ENTER] to cancel (current {self.rotational_step_size}): ")
-            if answer == "n":
-                return
-            try:
-                val = float(answer)
-            except ValueError:
-                pass
-        self.rotational_step_size = val
-        print(f"New step size set: {self.rotational_step_size}")
-
-    def _home(self):
-        confirm = input("Are you sure you want to home? Type 'yes' to confirm: ")
-        if confirm != "yes":
-            return
-        for semaphore in self.semaphores.values():
-            semaphore.acquire()
-        for motor in self.stage.motors:
-            if motor:
-                motor.home()
-        for semaphore in self.semaphores.values():
-            semaphore.release()
-        log.info("Homing complete")
-
-    def _help(self):
-        print(f"""
-        Stage Control
-        -------------
-        move left: {self.bindings.MOVE_LEFT}
-        move right: {self.bindings.MOVE_RIGHT}
-        move up: {self.bindings.MOVE_UP}
-        move down: {self.bindings.MOVE_DOWN}
-        jog left: {self.bindings.JOG_LEFT}
-        jog right: {self.bindings.JOG_RIGHT}
-        jog up: {self.bindings.JOG_UP}
-        jog down: {self.bindings.JOG_DOWN}
-        jog clockwise: {self.bindings.JOG_CLOCKWISE}
-        jog counterclockwise: {self.bindings.JOG_COUNTERCLOCKWISE}
-        linear jog step: {self.bindings.LINEAR_JOG_STEP}
-        rotational jog step: {self.bindings.ROTATIONAL_JOG_STEP}
-        home: {self.bindings.HOME}
-        help: {self.bindings.HELP}
-        quit: {self.bindings.QUIT}
-        """)
-        # stop all: {self.bindings.STOP_ALL}
-
-    def loop(self) -> None:
-        running = threading.Event()
-        running.set()
-
-        actions = list(self.bindings.dict().keys())
-        keys = list(self.bindings.dict().values())
-        funcs = {
-            "MOVE_LEFT": self._move_left,
-            "MOVE_RIGHT": self._move_right,
-            "MOVE_UP": self._move_up,
-            "MOVE_DOWN": self._move_down,
-            "JOG_LEFT": self._jog_left,
-            "JOG_RIGHT": self._jog_right,
-            "JOG_UP": self._jog_up,
-            "JOG_DOWN": self._jog_down,
-            "JOG_CLOCKWISE": self._jog_cw,
-            "JOG_COUNTERCLOCKWISE": self._jog_ccw,
-            "LINEAR_JOG_STEP": self._set_linear_jog_step,
-            "ROTATIONAL_JOG_STEP": self._set_rotational_jog_step,
-            "HOME": self._home,
-            "HELP": self._help,
-        }
-        flags = {binding : threading.Event() for binding in actions}
-
-        # keyboard.add_hotkey(self.bindings.MOVE_LEFT, lambda: self._move_left(semaphores[self.bindings.MOVE_LEFT]))
-        keyboard.add_hotkey(self.bindings.MOVE_LEFT, lambda: flags["MOVE_LEFT"].set())
-        keyboard.add_hotkey(self.bindings.MOVE_RIGHT, lambda: flags["MOVE_RIGHT"].set())
-        keyboard.add_hotkey(self.bindings.MOVE_UP, lambda: flags["MOVE_UP"].set())
-        keyboard.add_hotkey(self.bindings.MOVE_DOWN, lambda: flags["MOVE_DOWN"].set())
-        keyboard.add_hotkey(self.bindings.JOG_LEFT, lambda: flags["JOG_LEFT"].set())
-        keyboard.add_hotkey(self.bindings.JOG_RIGHT, lambda: flags["JOG_RIGHT"].set())
-        keyboard.add_hotkey(self.bindings.JOG_UP, lambda: flags["JOG_UP"].set())
-        keyboard.add_hotkey(self.bindings.JOG_DOWN, lambda: flags["JOG_DOWN"].set())
-        keyboard.add_hotkey(self.bindings.JOG_CLOCKWISE, lambda: flags["JOG_CLOCKWISE"].set())
-        keyboard.add_hotkey(self.bindings.JOG_COUNTERCLOCKWISE, lambda: flags["JOG_COUNTERCLOCKWISE"].set())
-        keyboard.add_hotkey(self.bindings.LINEAR_JOG_STEP, lambda: flags["LINEAR_JOG_STEP"].set())
-        keyboard.add_hotkey(self.bindings.ROTATIONAL_JOG_STEP, lambda: flags["ROTATIONAL_JOG_STEP"].set())
-        keyboard.add_hotkey(self.bindings.HOME, lambda: flags["HOME"].set())
-        keyboard.add_hotkey(self.bindings.HELP, lambda: flags["HELP"].set())
-        keyboard.add_hotkey(self.bindings.QUIT, lambda: running.clear())
-        keyboard.add_hotkey(self.bindings.HELP, lambda: flags["HELP"].set())
-
-        def run_flagged():
-            for action, flag in flags.items():
-                if flag.is_set():
-                    t = threading.Thread(target=funcs[action], daemon=True)
-                    t.start()
-                    flag.clear()
-
-        log.info("Entering keyboard control loop")
-        while running.is_set():
-            run_flagged()
-            time.sleep(0.1)
-        # else:
-        # clean up all current running actions, make sure all semaphores are freed
 
 
 def auto_calibration_callback(
@@ -286,7 +61,7 @@ def auto_calibration_callback(
     print("Optimizing alignment...")
     do_scan = True
     while do_scan:
-        auto_scan(stage, daq, SCAN_SPAN, SCAN_STEP, go_to_max=True)
+        auto_scan(stage, daq, SCAN_SPAN, SCAN_STEP)
         scan_again = input("Scan again? (y/n) [Enter]: ")
         if scan_again == "y":
             do_scan = True
@@ -321,6 +96,8 @@ def calibrate(
         corresponding order:
         - stage: The stage object to use for calibration.
         - daq: The data acquisition unit to use for calibration.
+        - circuit: The circuit currently being referenced for calibration.
+        - controller: The controller to use for control input.
 
     Returns
     -------
@@ -371,7 +148,7 @@ def calibrate(
 
     a = np.linalg.inv(gds) @ stage
 
-    conversion_matrix = np.array(
+    calibration_matrix = np.array(
         [
             [a[0][0], a[1][0], a[2][0]], 
             [a[3][0], a[4][0], a[5][0]], 
@@ -379,81 +156,7 @@ def calibrate(
         ]
     )
 
-    return conversion_matrix
-
-
-def easy_switcher(
-    stage: Stage, 
-    daq: DataAcquisitionUnitBase, 
-    callback: Callable = None, 
-    controller: KeyboardControl = None,
-) -> None:
-    """
-    Allows user to set specific locations for easy switching.
-
-    Can provide an optional callback that will be run after the user has moved
-    to a given location using the controller. It will be passed ``stage`` and
-    ``daq`` as parameters.
-
-    Switching circuits is done by pressing the associated key on the
-    keyboard. It is recommended to use one of the numeric keys on the
-    keyboard.
-
-    Switching circuits is only available in the context of this function; after
-    exiting the loop, location associations are discarded.
-
-    Parameters
-    ----------
-    stage : Stage
-        The stage device to use for interfacing with the hardware.
-    daq : DataAcquisitionUnitBase
-        The configured data acquisition unit (or class that inherits from
-        :py:class:`DataAcquisitionUnitBase`) to use for interfacing with the
-        hardware.
-    callback : Callable, optional
-        A callback function, could be used to optimize at each point. Could be 
-        an automatic scan or a function that interacts with the user. It should
-        block until the stage has settled on a point, but it does not need to
-        return anything; the settled point location will be obtained from the
-        Stage object. It should accept the following parameters, in the
-        corresponding order:
-        - stage: The stage object to use for calibration.
-        - daq: The data acquisition unit to use for calibration.
-    controller : KeyboardControl, optional
-        The controller to use for interfacing with the user. If not provided,
-        a default controller will be used.
-    """
-    again = True
-    locs = {}
-
-    if controller is None:
-        controller = KeyboardControl(stage)
-
-    while again:
-        print(f"Move to first desired location, then quit the controller.")
-        controller.loop()
-
-        if callback:
-            callback(stage, daq)
-
-        time.sleep(1.0)
-        print("Press the key you want to associate this point: ")
-        refkey = keyboard.read_hotkey()
-        locs[refkey] = [
-            stage.x.get_position(),
-            stage.y.get_position(),
-        ]
-        more = input("Log another point? (y, n): ")
-        if more != "y":
-            again = False
-
-    def go_to_position(x, y):
-        stage.set_position(x=x, y=y)
-
-    for key, loc in locs.items():
-        keyboard.add_hotkey(key, go_to_position, args=(loc[0], loc[1],))
-
-    controller.loop()
+    return calibration_matrix
 
 
 def basic_scan(
@@ -473,7 +176,6 @@ def basic_scan(
     step_size_y: float = None,
     plot: bool = False,
     settle: float = 0.2,
-    go_to_max: bool = True,
 ) -> Tuple[float, Tuple[float, float]]:
     """
     Performs a box scan of given dimensions and goes to position of highest readings returned.
@@ -554,9 +256,6 @@ def basic_scan(
         Amount of time in seconds the motors will pause in between move 
         commands to allow for settling time (lets vibration/residual motion 
         die out) and improve data reliability (default 0.2).
-    go_to_max : bool, optional
-        If True, the motors will be moved to the maximum position before
-        returning (default True).
 
     Returns
     -------
@@ -661,11 +360,11 @@ def basic_scan(
     max_x, max_y = pos[max_idx]
 
     # Move to max position
-    if go_to_max:
-        set_position_function(x=float(max_x), y=float(max_y))
+    set_position_function(x=float(max_x), y=float(max_y))
 
     if plot:
         plt.show()
+        print("Close the plot window to continue...")
 
     return max_data, (max_x, max_y)
 
@@ -762,6 +461,7 @@ def line_scan(
 
     if plot:
         plt.show()
+        print("Close the plot window to continue...")
 
     return max_loc
 
@@ -784,6 +484,8 @@ def auto_scan(
 ) -> Tuple[float, float]:
     """
     Performs a box scan of given dimensions and then refines the scan around the peak.
+
+    The stage always moves to the max position upon completion of this function.
 
     You must specify one of the following sets of parameters: 
     * ``stage_x`` and ``stage_y``
@@ -846,9 +548,6 @@ def auto_scan(
         Amount of time in seconds the motors will pause in between move 
         commands to allow for settling time (lets vibration/residual motion 
         die out) and improve data reliability (default 0.2).
-    go_to_max : bool, optional
-        If True, the motors will be moved to the maximum position before
-        returning (default True).
 
     Returns
     -------
